test: 💍 cover default case (split function from class)
