+++ src/base/Context.ts
@@ -7,21 +7,13 @@ import {
   PalletRelayerSubsidy,
   PolymeshCommonUtilitiesProtocolFeeProtocolOp,
 } from '@polkadot/types/lookup';
-import {
-  CallFunction,
-  Codec,
-  DetectCodec,
-  Signer as PolkadotSigner,
-  TypeDef,
-  TypeDefInfo,
-} from '@polkadot/types/types';
+import { CallFunction, Codec, DetectCodec, Signer as PolkadotSigner } from '@polkadot/types/types';
 import { SigningManager } from '@polymeshassociation/signing-manager-types';
 import { NormalizedCacheObject } from 'apollo-cache-inmemory';
 import ApolloClient, { ApolloQueryResult } from 'apollo-client';
 import BigNumber from 'bignumber.js';
 import P from 'bluebird';
 import { chunk, clone, flatMap, flatten, flattenDeep } from 'lodash';
-import { polymesh } from 'polymesh-types/definitions';
 
 import { Account, Asset, DividendDistribution, Identity, PolymeshError, Subsidy } from '~/internal';
 import { ClaimTypeEnum as MiddlewareV2Claim } from '~/middleware/enumsV2';
@@ -31,33 +23,23 @@ import { ClaimTypeEnum, Query } from '~/middleware/types';
 import { Query as QueryV2 } from '~/middleware/typesV2';
 import {
   AccountBalance,
-  ArrayTransactionArgument,
   ClaimData,
   ClaimType,
-  ComplexTransactionArgument,
   CorporateActionParams,
   DistributionWithDetails,
   ErrorCode,
   ModuleName,
-  PlainTransactionArgument,
   ProtocolFees,
   ResultSet,
-  SimpleEnumTransactionArgument,
   SubCallback,
   SubsidyWithAllowance,
   TransactionArgument,
-  TransactionArgumentType,
   TxTag,
   UnsubCallback,
 } from '~/types';
 import { GraphqlQuery } from '~/types/internal';
 import { Ensured, EnsuredV2 } from '~/types/utils';
-import {
-  DEFAULT_GQL_PAGE_SIZE,
-  MAX_CONCURRENT_REQUESTS,
-  MAX_PAGE_SIZE,
-  ROOT_TYPES,
-} from '~/utils/constants';
+import { DEFAULT_GQL_PAGE_SIZE, MAX_CONCURRENT_REQUESTS, MAX_PAGE_SIZE } from '~/utils/constants';
 import {
   accountIdToString,
   balanceToBigNumber,
@@ -90,6 +72,8 @@ import {
   delay,
   getApiAtBlock,
 } from '~/utils/internal';
+
+import { processType } from './utils';
 
 interface ConstructorParams {
   polymeshApi: ApiPromise;
@@ -664,135 +648,8 @@ export class Context {
        */
       _polymeshApi: { tx },
     } = this;
-    const { types } = polymesh;
 
     const [section, method] = tag.split('.');
-
-    const getRootType = (
-      type: string
-    ):
-      | PlainTransactionArgument
-      | ArrayTransactionArgument
-      | SimpleEnumTransactionArgument
-      | ComplexTransactionArgument => {
-      const rootType = ROOT_TYPES[type];
-
-      if (rootType) {
-        return {
-          type: rootType,
-        };
-      }
-      if (type === 'Null') {
-        return {
-          type: TransactionArgumentType.Null,
-        };
-      }
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const definition = (types as any)[type];
-
-      if (!definition) {
-        return {
-          type: TransactionArgumentType.Unknown,
-        };
-      }
-
-      const typeDef = getTypeDef(JSON.stringify(definition));
-
-      if (typeDef.info === TypeDefInfo.Plain) {
-        return getRootType(definition);
-      }
-
-      // eslint-disable-next-line @typescript-eslint/no-use-before-define
-      return processType(typeDef, '');
-    };
-
-    const processType = (rawType: TypeDef, name: string): TransactionArgument => {
-      const { type, info, sub } = rawType;
-
-      const arg = {
-        name,
-        optional: false,
-        _rawType: rawType,
-      };
-
-      switch (info) {
-        case TypeDefInfo.Plain: {
-          return {
-            ...getRootType(type),
-            ...arg,
-          };
-        }
-        case TypeDefInfo.Compact: {
-          return {
-            ...processType(sub as TypeDef, name),
-            ...arg,
-          };
-        }
-        case TypeDefInfo.Option: {
-          return {
-            ...processType(sub as TypeDef, name),
-            ...arg,
-            optional: true,
-          };
-        }
-        case TypeDefInfo.Tuple: {
-          return {
-            type: TransactionArgumentType.Tuple,
-            ...arg,
-            internal: (sub as TypeDef[]).map((def, index) => processType(def, `${index}`)),
-          };
-        }
-        case TypeDefInfo.Vec: {
-          return {
-            type: TransactionArgumentType.Array,
-            ...arg,
-            internal: processType(sub as TypeDef, ''),
-          };
-        }
-        case TypeDefInfo.VecFixed: {
-          return {
-            type: TransactionArgumentType.Text,
-            ...arg,
-          };
-        }
-        case TypeDefInfo.Enum: {
-          const subTypes = sub as TypeDef[];
-
-          const isSimple = subTypes.every(({ type: subType }) => subType === 'Null');
-
-          if (isSimple) {
-            return {
-              type: TransactionArgumentType.SimpleEnum,
-              ...arg,
-              // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-              internal: subTypes.map(({ name: subName }) => subName!),
-            };
-          }
-
-          return {
-            type: TransactionArgumentType.RichEnum,
-            ...arg,
-            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-            internal: subTypes.map(def => processType(def, def.name!)),
-          };
-        }
-        case TypeDefInfo.Struct: {
-          return {
-            type: TransactionArgumentType.Object,
-            ...arg,
-            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-            internal: (sub as TypeDef[]).map(def => processType(def, def.name!)),
-          };
-        }
-        default: {
-          return {
-            type: TransactionArgumentType.Unknown,
-            ...arg,
-          };
-        }
-      }
-    };
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     return ((tx as any)[section][method] as CallFunction).meta.args.map(({ name, type }) => {
